Moved general channel tests to channels module
