+++ Tribler/Test/Core/Modules/RestApi/Channels/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains tests for the channels endpoints.
+"""
